refactor file structure change
