+++ src/main/java/com/example/onboarding/dto/LoginRequestDto.java
@@ -1,0 +1,13 @@
+package com.example.onboarding.dto;
+
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class LoginRequestDto {
+    private String email;
+    private String password;
+}
