+++ src/main/java/com/example/onboarding/dto/MissionListDto.java
@@ -1,0 +1,18 @@
+package com.example.onboarding.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class MissionListDto {
+    private String title;
+    private LocalDateTime deadline;
+
+}
