+++ src/main/java/com/example/onboarding/dto/MissionResponseDto.java
@@ -1,0 +1,32 @@
+package com.example.onboarding.dto;
+
+import com.example.onboarding.entity.Mission;
+import lombok.*;
+
+import java.time.LocalDateTime;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class MissionResponseDto {
+
+    private int missionId;
+    private String title;
+    private String description;
+    private LocalDateTime deadline;
+    private int totalGroups;
+    private LocalDateTime createdAt;
+
+    public static MissionResponseDto fromEntity(Mission mission) {
+        return MissionResponseDto.builder()
+                .missionId(mission.getMissionId())
+                .title(mission.getTitle())
+                .description(mission.getDescription())
+                .deadline(mission.getDeadline())
+                .totalGroups(mission.getTotalGroups())
+                .createdAt(mission.getCreatedAt())
+                .build();
+    }
+}
