+++ src/main/java/com/example/onboarding/dto/ReplyRequest.java
@@ -1,0 +1,12 @@
+package com.example.onboarding.dto;
+
+import lombok.*;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class ReplyRequest {
+    private int userId;
+    private int groupId;
+    private String content;
+}
