+++ src/main/java/com/example/onboarding/dto/ReplyResponse.java
@@ -1,0 +1,20 @@
+package com.example.onboarding.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class ReplyResponse {
+    private int replyId;
+    private int userId;
+    private int groupId;
+    private String content;
+    private LocalDateTime finalTime;
+}
