+++ src/main/java/com/example/onboarding/dto/UserRequestDto.java
@@ -1,0 +1,15 @@
+package com.example.onboarding.dto;
+
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class UserRequestDto {
+  private String nickname;
+  private String email;
+  private String password;
+  private boolean isAdmin;
+}
