+++ src/main/java/com/example/onboarding/dto/UserResponseDto.java
@@ -1,0 +1,14 @@
+package com.example.onboarding.dto;
+
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class UserResponseDto {
+    private String email;
+    private String nickname;
+    private boolean isAdmin;
+}
