+++ src/main/java/com/example/onboarding/entity/GroupEntity.java
@@ -1,0 +1,52 @@
+package com.example.onboarding.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+
+@Entity
+@Table(name = "`group`")
+public class GroupEntity {
+	
+	@Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "group_id")
+    private int groupId;
+
+    @ManyToOne(optional = false)
+    @JoinColumn(name = "mission_id", nullable = false)
+    private Mission missionId;
+
+    @Column(name = "group_num", nullable = false)
+    private int groupNum;
+
+	public int getGroupId() {
+		return groupId;
+	}
+
+	public void setGroupId(int groupId) {
+		this.groupId = groupId;
+	}
+
+	public Mission getMissionId() {
+		return missionId;
+	}
+
+	public void setMissionId(Mission missionId) {
+		this.missionId = missionId;
+	}
+
+	public int getGroupNum() {
+		return groupNum;
+	}
+
+	public void setGroupNum(int groupNum) {
+		this.groupNum = groupNum;
+	}	
+    
+}
