--- src/main/java/com/example/onboarding/mission/dto/MissionListDto.java
@@ -1,19 +1,0 @@
-package com.example.onboarding.mission.dto;
-
-import com.example.onboarding.mission.entity.Mission;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class MissionListDto {
-    private String title;
-    private LocalDateTime deadline;
-
-}
