--- src/main/java/com/example/onboarding/mission/entity/Mission.java
@@ -1,35 +1,0 @@
-package com.example.onboarding.mission.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "mission")
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class Mission {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "mission_id")
-    private Integer missionId;
-
-    @Column(nullable = false, length = 100)
-    private String title;
-
-    @Column(columnDefinition = "TEXT")
-    private String description;
-
-    @Column(nullable = false)
-    private LocalDateTime deadline;
-
-    @Column(name = "total_groups", nullable = false)
-    private int totalGroups;
-
-    @Column(name = "created_at", nullable = false)
-    private LocalDateTime createdAt;
-}
