--- src/main/java/com/example/onboarding/missionWrite/repository/GroupRepository.java
@@ -1,9 +1,0 @@
-package com.example.onboarding.missionWrite.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.example.onboarding.missionWrite.Entity.GroupEntity;
-
-public interface GroupRepository extends JpaRepository<GroupEntity, Integer> {
-
-}
