--- src/main/java/com/example/onboarding/reply/model/ReplyRequest.java
@@ -1,12 +1,0 @@
-package com.example.onboarding.reply.model;
-
-import lombok.*;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class ReplyRequest {
-    private int userId;
-    private int groupId;
-    private String content;
-}
