--- src/main/java/com/example/onboarding/reply/repository/ReplyRepository.java
@@ -1,7 +1,0 @@
-package com.example.onboarding.reply.repository;
-
-import com.example.onboarding.reply.entity.Reply;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ReplyRepository extends JpaRepository<Reply, Integer> {
-}
