+++ src/main/java/com/example/onboarding/repository/GroupRepository.java
@@ -1,0 +1,9 @@
+package com.example.onboarding.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.example.onboarding.entity.GroupEntity;
+
+public interface GroupRepository extends JpaRepository<GroupEntity, Integer> {
+
+}
