+++ src/main/java/com/example/onboarding/repository/ReplyRepository.java
@@ -1,0 +1,7 @@
+package com.example.onboarding.repository;
+
+import com.example.onboarding.entity.Reply;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ReplyRepository extends JpaRepository<Reply, Integer> {
+}
