+++ src/main/java/com/example/onboarding/repository/UserRepository.java
@@ -1,0 +1,11 @@
+package com.example.onboarding.repository;
+
+import com.example.onboarding.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Integer> {
+    Optional<User> findByEmail(String email);
+    boolean existsByEmail(String email);
+}
