--- src/main/java/com/example/onboarding/user/dto/LoginRequestDto.java
@@ -1,13 +1,0 @@
-package com.example.onboarding.user.dto;
-
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class LoginRequestDto {
-    private String email;
-    private String password;
-}
