--- src/main/java/com/example/onboarding/user/dto/UserRequestDto.java
@@ -1,15 +1,0 @@
-package com.example.onboarding.user.dto;
-
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class UserRequestDto {
-  private String nickname;
-  private String email;
-  private String password;
-  private boolean isAdmin;
-}
