--- src/main/java/com/example/onboarding/user/dto/UserResponseDto.java
@@ -1,14 +1,0 @@
-package com.example.onboarding.user.dto;
-
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class UserResponseDto {
-    private String email;
-    private String nickname;
-    private boolean isAdmin;
-}
