--- src/main/java/com/example/onboarding/user/entity/User.java
@@ -1,28 +1,0 @@
-package com.example.onboarding.user.entity;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Table(name = "user")
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int userId;
-
-    @Column(nullable = false, length = 100)
-    private String nickname;
-
-    @Column(nullable = false, unique = true, length = 100)
-    private String email;
-
-    @Column(nullable = false)
-    private String password;
-
-    @Column(nullable = false)
-    private boolean isAdmin;
-}
