--- src/main/java/com/example/onboarding/user/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package com.example.onboarding.user.repository;
-
-import com.example.onboarding.user.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Integer> {
-    Optional<User> findByEmail(String email);
-    boolean existsByEmail(String email);
-}
